requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.3
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1
